corrected schema
