+++ routes/api/contacts.js
@@ -3,27 +3,27 @@ const ctrl = require("../../controllers/contacts");
 const router = express.Router();
 
 const { validateBody, isValidId } = require("../../middlewares");
-const schema = require("../../schemas/contacts");
+const { schemas } = require("../../models/contact");
 
 router.get("/", ctrl.listContacts);
 
 router.get("/:contactId", isValidId, ctrl.getContactById);
 
-router.post("/", validateBody(schema.addSchema), ctrl.addContact);
+router.post("/", validateBody(schemas.addSchema), ctrl.addContact);
 
 router.delete("/:contactId", isValidId, ctrl.removeContact);
 
 router.put(
   "/:contactId",
   isValidId,
-  validateBody(schema.addSchema),
+  validateBody(schemas.addSchema),
   ctrl.updateContact
 );
 
 router.patch(
   "/:contactId/favorite",
   isValidId,
-  validateBody(schema.updateFavoriteSchema),
+  validateBody(schemas.updateFavoriteSchema),
   ctrl.updateFavorite
 );
