--- schemas/contacts.js
@@ -1,17 +1,0 @@
-const Joi = require("joi");
-
-const addSchema = Joi.object({
-  name: Joi.string().required(),
-  email: Joi.string().email().required(),
-  phone: Joi.string().min(3).max(20).required(),
-  favorite: Joi.boolean,
-});
-
-const updateFavoriteSchema = Joi.object({
-  favorite: Joi.boolean().required(),
-});
-
-module.exports = {
-  addSchema,
-  updateFavoriteSchema,
-};
